Add UserIDView to authentication views and URLs
user id can optain with a get request to this endpoint with given credentials

The new UserIDView has been added to the authentication views. It makes use of the Django Rest Framework's APIView, permission classes, and response. Furthermore, a new URL pattern for the UserIDView has been added to the urlpatterns in the authentication's URLs file.