Add 'common' app and expand user details
A new Django app, 'common', has been added to manage documents associated with users. Fields 'first_name', 'last_name', and 'date_of_birth' have been added to the user app's model 'PersonalDetails', and the 'identity' field has been removed. This refines personalization and supports the introduction of the 'common' app. The models of this new app include 'Document' and 'Photo', associated with a user and allowing the user to have related documents and photos saved in the system.