Enable hive view set and add document fields to models
The HiveViewSet has been enabled in the urls.py file to allow for the handling of hive related requests. Moreover, a "documents" field has been added to the Hive, Bee, Gig, HiveRequest, and Contract models in the models.py file. This field links each model to common documents, facilitating document sharing and handling in the system.