Add admin permissions check to Hive methods
Added new methods in the Hive model to check if a user or bee is an admin. These checks have been applied in the views for update, delete and partial update actions to prevent non-admins from performing these operations on a Hive object. This implementation enhances the security and data integrity of the application.