Add communication module to app
A new communication module was created. It includes models for conversation and messages. The use of serializers allows responses to be converted into 'python native datatypes' that can then be rendered into 'json' - allowing for API functionality. Permissions were added to ensure user and public hive accessibility. Communication signals have also been established to sync conversation participants. Finally, new views support the expansion of the project and the urls help in routing. This module is tied into the djangoProject settings. Modifications were also made in the honeycomb models to accommodate the communication module.