Update task assignment and feedback filtering logic
Updated the way tasks are reassigned and feedbacks are filtered. Now, tasks are directly assigned instead of being reassigned and feedbacks are filtered based on the author instead of the user. Further, if a task is not searching for bees, all unaccepted assignments for that task are deleted. Implemented an 'accept' action in viewsets to allow accepting the assignment with proper validation and response handling.