Update accept_application method to handle hive_request
The method accept_application in honeycomb's models has been updated to accept an additional parameter "hive_request_id". This is used to fetch the corresponding hive request from the HiveRequest table and set its status to accepted, replicating the behavior done for a bee's membership. The call of this method in honeycomb's views has been updated correspondingly.