Add task details to MembershipSerializer and update MembershipAdmin
In MembershipSerializer, added two new fields to fetch and serialize the tasks assigned to a particular member and their task assignments. In MembershipAdmin, we updated the displayed and filterable fields, eliminating 'joined_at' from both the display and filter options and also excluding it from the admin edit fields, to streamline the interface.