Add Skill related view, route, serializer, and filter
The commit adds a view for Skill model in viewsets and a corresponding route in the router. It also introduces a serializer for the Skill model to allow its data to be serialized into a format that can be rendered into JSON. Finally, a filter is added for the Skill model to perform advanced filtering based on the model's fields.