Refactor user model and update serializers
This commit includes updates to the User model to ensure the existence of 'personal_details' and allow setting 'first_name' and 'last_name'. It introduces getters and setters for these properties. The serializers have been updated to support these changes. In addition, the 'fields' property of the user serializer in HelpDesk has been set to "__all__" to include all fields.