Add hashing and document re-use to Document model
The commit introduces hashing for document files and checks for the existence of the same hash before saving a new document. If the same hash exists, the current instance of document will reuse the existing document's details. The save method in Document model has been updated accordingly and a new method to calculate the hash has been added.